renamed query to api, renamed stats to db
