+++ api/api.go
@@ -1,4 +1,4 @@
-package query
+package api
 
 import (
 	"bytes"
